+++ app/NEWS.md
@@ -5,6 +5,7 @@
 ### new features / performance tweaks / improvements
 
 - drawing zero line in OSCI again
+- added webkitOfflineAudioContext to new wav parser routine. This should make the webApp work with common sample rates on Safari (only 44100, 4800 and 96000 seems to work though)
 
 ### bug fixes